Reorder logic in assigning indices to variables
There was an edge case in the previous logic. If a variable did not match its conjugate, and if the conjugate had been seen, they would be assigned different variables. This commit fixes the problem, by assigning to the variable that of the conjugate if the conjugate is present already.

THIS HAS PHYSICAL IMPLICATIONS: Using the same configurations as arXiv:1909.10519, the upper bounds to the Svetlichny inequality and the Mermin inequality are 4 and 2.006, respectively. For this reason, we must check very carefully that the change is correct.