Use commutation of disconnected components for canonical form
The two-body commutation rules in self.substitutions are not enough in some occasions when the monomial can be factorized. An example is (all indices are inflation indices) A13A33A22 and A22A13A33. Both are the same because they are composed of two disconnected monomials, but cannot be brought one to another via two-body substitution rules. This commit fixes this issue, and treats both monomials as the same. This partially addresses #39, in the sense that the number of columns is independent of the order of the parties with this fix, but yet the generated moment matrices have a different number of variables.

This fix required making a copy of factorize_monomial() to fast_npa.py in order to overcome a circular import error. Ideally this will not be needed in the future. Also, in to_canonical and to_representative_aux the support for JIT had to be removed.