Use dictionary of semiknown relations
In order to accommodate partial specifications of known moments (in contrast with full probability distributions), this commit transforms the list semiknown_moments and substitutes it by a dictionary where the keys correspond to the variables that can be associated LPI constraints (see, e.g., arXiv:2203.16543, Eq. D6) and the values are a list with the variable it is proportional to and the proportionality constant.

Also, removed commented, unused code in sdp_utils.