Prepare set_distribution to accept atomic values
We add a new argument to set_values, only_specified_values. The goal of this flag will be to just parse the input to self.known_moments if True, but if False it will assign values to variables representing products of variables that are assigned numerical values in the input.