Rework processing of certificates
Now the infeasibility certificates read and process the output of the solver function directly. This change includes  the development of a new function certificate_as_string, reworking certificate_as_probs and certificate_as_objective, and modifying the function clean_coefficients for handling the new specification of the certificate. For the time being, the function certificate_as_correlators is not implemented.


Former-commit-id: a0dd2c7911a981aa630af4b341c97d684a49b5c6