Remove _update_objective
The way of updating the objective after setting known values did not allow to properly update it if new known values were set. In order to solve this, now if an objective exists it is completely processed every time new values are set.

The remaining task of passing the semiknowns through the objective now is moved to set_objective.