Remove positive_vars from input to the solver
Because the generation of the problem sets the corresponding var_lowerbounds for all positive variables, we do not need to feed both var_lowerbounds and positive_vars to the solver. Moreover, it is not good practice to modify the input problem (i.e., overriding the lower bounds) in the function whose task is only to write the problem in a solver-compatible form and solve it. If there are any discrepancies between var_lowerbounds and positive_vars, these should be reconciled in InflationSDP.


Former-commit-id: b1e6cd3347f31c6fb775889a03a4cf7d74c7b120