Changed functions to admit custom lexicographic ordering, encoded as lexorder=np.array([op_A,op_B,op_C,...]). The hash of an operator is its position in lexorder. e.g. hash(op_B) = 1, hash(op_A) = 0, etc.
Currently all tests pass.

The main advantage of this is that this hash can serve as indices to lookup in a matrix. for example, to find if op1 commutes with op2, we can look this up in a commutation matrix COMM[hash(op1),hash(op2)]. The COMM matrix (TO BE IMPLEMENTED NEXT) is computed at the beginning and it can have encode any possible commutation relationships.

Furthermore, it is interesting to be flexible with lexicographic ordering for the identification of zero unknowable monomials that I mentioned in the email, when some support is 0, some unknowable monomials can also be 0, but the 0 operator can be hidden inside, but it can be brought to the surface by changing the lexicographic ordering.


Former-commit-id: afd998682afdb5aba806b2e8bbbcb6327d89f702