Add test for _sanitize_monomial
I tried to cover all the relevant cases, but please add those I could have missed.

Note that tests 4 and 6 currently fail. This is because the input is not written in canonical form. If the expected behavior of _sanitize_monomial is not bringing the monomials to canonical form, test 4 shall be fixed and test 6 shall be removed.