Fix example on optimization of CHSH
The certificate obtained that recovers the CHSH inequality was being interpreted incorrectly. The certificate itself is the inequality. The witness of infeasibility is cert < 0, which upon rewriting gives the CHSH inequality with its Tsirelson bound. When maximizing the quantity, we were looking in the wrong direction, because infeasibility is witnessed by negative values. Minimizing instead gives something close to the expected certificate, cert > 0 for feasible distributions.

This commit changes the example in the notebook to illustrate the correct way of proceeding