Added method to InflationSDP to convert monomials to symbol. Works with various input types. Added tests for it.
Modified the Mermin example to return symbolic columns. Reran the example so the notebook shows the correct number.

Numba doesnt allow for raising exceptions, so added a short print message at the end of nb_lexorder_idx if the operator is not found in the matrix (I believe there is no performance loss as the print statement is never reached if the input is good). I lost a bit of time debugging this, and it might be a common error in the future.

_interpret_atomic_string cannot actually process conditional probabilities, so removed outdated error message.