Make moment matrix progress bar progress linearly
Tqdm when calculating the moment matrix is quite useless in predicting the total time it will take to compute. This is because we call tqdm on the row index, but as we are advancing in the computation, the number of columns we have to loop through decreases, thus the progress bar speeds up.

With this change, I've found that the estimated remaining time is fairly accurate! This is quite useful.