Completed missing features for writing to .mat files.
Add export to matlab of inequalities (only equalities were considered).

Made so that the exported object is read as a struct in MATLAB making it also easier to interpret. E.g. in matlab now it would be moment_inequality{5}.moments is a list of indexes, and moment_inequality{5}.coeffs is a list of coeffs corresponding to the 5th inequality. This makes the data structure constant. before if all inequalities have the same number of terms it would be n x m x 2 array, but when there would be a different number of terms for one inequality, it would become a cell. This would be annoying for a MATLAB user if they have to change code for different problems.

For bounds, I use the more up to date _processed_**.

Change the names to coincide (more or less) with those of the Class attributes.