Added test for the processing of eqs and ineqs with lpi constraints.
Discovered a bug where the input to solveSDP was begin modified, giving erroneous results over multiple calls to the function. Given the extremely sparse nature of the objective, inequalities and equalities, doing an internal copy of these shouldn't have a big overhead. For now, when initialising the internal dictionaries, I simply copy the inputs dicts (except for the mask matrices!).

I made the input be named "objective" "inequalities" "equalities" and the internal copies  "var_objective" "var_inequalities", "var_equalities"

When processing semiknown constraints I noticed for LP problems sometimes Fi would not be empty!! This is because with Fi[x2] = Fi.get(x2, 0) + c * val it would add x2 even if Fi is empty! I added an "if mask_matrices" to fix this. In principle the same could happen to "objective" so I add an "if var_objective" (taking into account that the objective could be non-empty for feas_as_optim problems)

Indentation on a line