Adding the option 'full_expansion' to `expand_moment_normalisation``to allow for the expansion of moments such as A1B1 into a single constraint 1-A0-B0+A0B0 as opposed to two constraints, (B1-A0B1, A1-A0B1).
Each of these expressions will be encoded as (lhs,rhs).
Eg:
A1B10=1-A0-B0+A0B0 -> ((A1B1,A0,B0),(1,A0B0)).
One can think of "lhs" as all the positive terms, and "rhs" all the negative terms.

Also added a flag to allow for omitting the input monomial, in the above A1B1, from the resulting expression. This is relevant if we want to use the output to do substitution, and not to use it as a constraint.

Also added some simple tests.