Modified InflationProblem to have an attribute _quantum_sources. This indicates which of the sources are quantum and which are not. Also added relevant attributes to InflationSDP to take this into account.
The changes in the code are the following: when checking overlapping support with inflation indices, the columns of all classical sources are removed. Then if ALL columns are removed -all sources classical- we say all operators commute. TECHNICALLY, even if all sources are classical, for different inputs technically speaking an operators doesn't need to commute, but given that there is no way to take advantage of non-commuting measurements for nonlocality with all classical sources, there is no need to consider this.