Fix test corresponding to 9694799
The specification of columns [0, 0] (and analogous ones) should only generate the monomials of type [0, 0], and not those of type [0] as well. That is the point of having this level of granularity, instead of, for instance, just specifying the number of operators of each party. In fact, this commit ant 9694799 are just recovering the behavior that was present in v1.0.0