Initial support for sparse internal operations
There is an internal memory bottleneck when using the function nb_outer_bitwise_or. This is an operation on boolean arrays to boolean arrays, so it can greatly benefit from rewriting it as sparse operations. This commit is an initial implementation of this.

Due to the fact that we need to do broadcasting operations to tensors, I am using the PyData sparse library.

Some tests are failing, due to the necessity of rewriting the parts with vstack when handling equalities. Nevertheless, the basic functionality works.

However, the code suffers from the absence of JIT compilation